+++ jax/random.py
@@ -170,17 +170,17 @@ from jax._src.random import (
   randint as randint,
   random_gamma_p as random_gamma_p,
   rayleigh as rayleigh,
-  rbg_key as rbg_key,
+  rbg_key as _deprecated_rbg_key,
   shuffle as shuffle,
   split as split,
   t as t,
   threefry_2x32 as threefry_2x32,
-  threefry2x32_key as threefry2x32_key,
+  threefry2x32_key as _deprecated_threefry2x32_key,
   threefry2x32_p as threefry2x32_p,
   triangular as triangular,
   truncated_normal as truncated_normal,
   uniform as uniform,
-  unsafe_rbg_key as unsafe_rbg_key,
+  unsafe_rbg_key as _deprecated_unsafe_rbg_key,
   wald as wald,
   weibull_min as weibull_min,
   wrap_key_data as wrap_key_data,
@@ -202,12 +202,25 @@ _deprecations = {
         "jax.dtypes.issubdtype(arr, jax.dtypes.prng_key) for runtime detection of "
         "typed prng keys.", _PRNGKeyArray
     ),
+    # Added September 21, 2023
+    "threefry2x32_key": (
+        "jax.random.threefry2x32_key(seed) is deprecated. "
+        "Use jax.random.PRNGKey(seed, 'threefry2x32')", _deprecated_threefry2x32_key),
+    "rbg_key": (
+        "jax.random.rbg_key(seed) is deprecated. "
+        "Use jax.random.PRNGKey(seed, 'rbg')", _deprecated_rbg_key),
+    "unsafe_rbg_key": (
+        "jax.random.unsafe_rbg_key(seed) is deprecated. "
+        "Use jax.random.PRNGKey(seed, 'unsafe_rbg')", _deprecated_unsafe_rbg_key),
 }
 
 import typing
 if typing.TYPE_CHECKING:
   PRNGKeyArray = typing.Any
   KeyArray = typing.Any
+  threefry_2x32_key = _deprecated_threefry2x32_key
+  rbg_key = _deprecated_rbg_key
+  unsafe_rbg_key = _deprecated_unsafe_rbg_key
 else:
   from jax._src.deprecations import deprecation_getattr as _deprecation_getattr
   __getattr__ = _deprecation_getattr(__name__, _deprecations)